Update patch set 2
Patch Set 2: Code-Review+1

Thanks for the quick reply. Yes, I did verify it compiles, runs and behaves as expected.
I don't have the selection of verified flag when reply, and my Code-Review range is from -1 to 1. My guess is that I don't have the permission to do so.
Based on my reading maybe you could add "Change Owner" group to Reference > refs/for/refs/* > Label Verified?
PS: the manual says in the reviewer list, "for each reviewer there is a tooltip that shows on which labels the reviewer is allowed to vote", which I can't see either. My tooltip only shows name and email.

Patch-set: 2