Update patch set 5
Patch Set 5:

xmo, thank you for the contribution!
I hope it is not the last one ;)

note that the patch wasn't merged automatically to github (I'm not sure why though) so I did it manually, you might notice that the commit message changed. I'm still checking GerritHub to see if that's the right tool to use, we'll see..

Patch-set: 5