plugin: container images registry
Added a plugin for connecting to an image registry server.
The plugin adds a tab to 'Connect to server' dialog:
* Server: e.g. docker.io
* Image: e.g. cirros
* Type: docker, oci or local dir
* Credentials (if required)

Upon connection 'skopeo inspect' is invoked to get the
image's layers (each layer is a .tar.gz file).
Then, 'skopeo copy' is invoked to pull the image into
a temp directory.