eliminate a race that hinders automatic refresh
There was a potential race that may have caused changes in currently
presented folders to get unnoticed by the automatic refreshs mechanism:
1. The location of a folder panel is changed
2. The children of the presented folder are listed
3. The children of the presented folder are changed
4. The modification date of the presented folders is queried
5. Consequently, the presented folder wouldn't get refreshed and so
   the changes in step 4 wouldn't have been reflected in the UI.

This is solved by querying the folder's modification date prior to
listing its children. This may result in a redundant refresh later
on, which is better than the possibily of the UI not catching up
with changes to the currently presented folder.