fix #379: default move to trash to Apple-script on macOS
The reason for changing the implementation of the move-to-trash
operation on macOS to be based on JNA was to workaround issues
that may happen with the security and private settings on macOS
that prevent us from interacting with Finder.

The first problem with this approach that was identified is that
JNA doesn't move files on SMB shares to the SAMBA trash folder.
That was fixed already. But now it also appears that on newer
(beta) versions of macOS, JNA fails to move all files to trash
whereas it works properly with Apple script.

So it seems that a better approach on macOS would be to try to move
files to trash using an Apple-script first and only if that fails
to fall back to JNA.