change the bundled jre on macos back to java 8
We originally switched to Java 11 due to issues with macOS-features (setting
'about' and 'preferences' dialogs) that were observed with Java 8 on newer
versions of macOS.

Java 11.0.7 was reported as unstable and thus we reverted back to 11.0.4.
However, as a new version of Java 8 was recently released in which the
above mentioned issues are fixed and as Java 8 is the latest version that
is available for older versions of macOS (older than 10.10), it makes more
sense to switch to it (version 1.8.0_265) instead.