fix #695: improve rendering of editable filename cell
Previously, as part of the fix for #515, the text field became sort
of hidden by setting its colors to those of the selected row in the
file table. That confused users and made them believe that the cell
is not editable anymore when there was no selected text.

Now, the text field is set with its default colors and with a line
border rather than an empty border.