fix flakiness of AppleScriptTest#testScriptOutput
the test compares the received output against the expected output "6"
however, the 'osascript' command returns a new line at the end of the
generated output that we trim on ScriptOutputListener#processDied. it
could be that ScriptOutputListener#processDied would run after the
above mentioned comparison and then we get: "6" != "6\n".

we can wait for ScriptOutputListener#processDied to be completed
before handling the output, but we already trim the output to resolve
this issue elsewhere (in OSXTrash#getItemCount) so we rather apply
the same approach to AppleScriptTest#testScriptOutput.