nightly: create dmg with Java 18
There is an issue in creating a dmg from a signed app-image with Java 20,
apparently the signature of the files in app-image is not preserved when
they are copied to the folder from which the dmg is created. As a temporary
solution, switching the version of Java that is used to compile the source
and creating the dmg to 18 instead of 20.